# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GameResult {
  id: number;
  /** Canonical (lowercased) player name — used as the key for password reset. */
  playerName: string;
  firstName: string;
  lastName: string;
  score: number;
  rank: string;
  /** @nullable */
  timeC1: number | null;
  /** @nullable */
  timeC2: number | null;
  /** @nullable */
  timeC3: number | null;
  /** @nullable */
  timeC4: number | null;
  /** @nullable */
  totalTime: number | null;
  isTest: boolean;
  createdAt: string;
}

export interface GameResultInput {
  /**
     * The canonical player name (lowercased) used to look up the player record for authentication.
     * @minLength 1
     */
  playerName: string;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minimum 0 */
  score: number;
  /** @minLength 1 */
  rank: string;
  /** @nullable */
  timeC1: number | null;
  /** @nullable */
  timeC2: number | null;
  /** @nullable */
  timeC3: number | null;
  /** @nullable */
  timeC4: number | null;
  /** @nullable */
  totalTime: number | null;
  isTest: boolean;
}

export interface ResultsSummary {
  participantCount: number;
  averageScore: number;
  /** @nullable */
  topScore: number | null;
  /** @nullable */
  averageTime: number | null;
  /** @nullable */
  fastestTime: number | null;
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  /**
     * Unique username (login key, lowercase-normalised).
     * @minLength 1
     */
  name: string;
  /** Full display name shown on the leaderboard (optional on login, required on first registration). */
  displayName?: string;
  /** @minLength 1 */
  password: string;
}

export type LoginResultTimes = { [key: string]: unknown };

export interface LoginResult {
  isNew: boolean;
  name: string;
  displayName: string;
  score: number;
  won: number[];
  times: LoginResultTimes;
  submitted: boolean;
  updatedAt: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResult {
  authenticated: boolean;
}

export interface AdminSessionResult {
  authenticated: boolean;
}

export interface ResetPasswordInput {
  /**
     * Username of the account to reset.
     * @minLength 1
     */
  name: string;
  /** @minLength 1 */
  password: string;
}

export type PlayerProgressTimes = { [key: string]: unknown };

export interface PlayerProgress {
  name: string;
  score: number;
  won: number[];
  times: PlayerProgressTimes;
  submitted: boolean;
  updatedAt: string;
}

export interface AdminAuthInput {
  /**
     * The organizer passcode — validated against SESSION_SECRET on the server.
     * @minLength 1
     */
  passcode: string;
}

export interface AdminAuthResult {
  ok: boolean;
}

export interface AdminResetInput {
  /**
     * The organizer passcode — validated against SESSION_SECRET on the server.
     * @minLength 1
     */
  adminPasscode: string;
}

export type PasswordResetResultTimes = { [key: string]: unknown };

/**
 * Returned when an admin-initiated password reset succeeds. Contains the one-time token to share with the player.
 */
export interface PasswordResetResult {
  name: string;
  score: number;
  won: number[];
  times: PasswordResetResultTimes;
  submitted: boolean;
  updatedAt: string;
  /** One-time reset token — share this with the player. Valid for 24 hours. */
  resetToken: string;
  /** ISO timestamp when the reset token expires. */
  expiresAt: string;
}

export interface LoginResetInput {
  /** @minLength 1 */
  name: string;
  /**
     * The one-time reset token issued by the admin.
     * @minLength 1
     */
  resetToken: string;
  /** @minLength 1 */
  password: string;
}

export type PlayerProgressInputTimes = { [key: string]: unknown };

export interface PlayerProgressInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  score: number;
  won: number[];
  times: PlayerProgressInputTimes;
  submitted: boolean;
}