# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: results
    description: Cyber Treasure Hunt participant results
  - name: progress
    description: Player progress persistence
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List leaderboard results
      description: Returns participant results ordered by competition score.
      responses:
        "200":
          description: Leaderboard results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameResult"
    post:
      operationId: submitResult
      tags: [results]
      summary: Submit or update a participant result
      description: Upserts a participant's latest completed game result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameResultInput"
      responses:
        "200":
          description: Saved participant result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResult"
        "400":
          description: Invalid result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      operationId: login
      tags: [progress]
      summary: Create account or login
      description: Creates a new player account or verifies an existing one's password. Returns saved progress on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [progress]
      summary: Sign in to the organizer dashboard
      description: Verifies an organizer username and password and starts an admin session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid admin credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [progress]
      summary: Check the organizer session
      description: Confirms that the current browser has an active organizer session.
      responses:
        "200":
          description: Admin session is active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionResult"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reset-password:
    post:
      operationId: resetPassword
      tags: [progress]
      summary: Reset a player's password by username
      description: Allows a player to set a new password using their username alone. No second factor — appropriate for an internal game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Username not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/{name}:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get player progress
      description: Returns the saved progress for a player by name. Returns 404 if not found.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProgress"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: saveProgress
      tags: [progress]
      summary: Save player progress
      description: Upserts progress for a player.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerProgressInput"
      responses:
        "200":
          description: Saved progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProgress"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth:
    post:
      operationId: verifyAdminPasscode
      tags: [progress]
      summary: Verify admin passcode
      description: Validates the supplied passcode against the server-configured SESSION_SECRET. Returns 200 on success, 401 on wrong passcode, 503 if admin auth is not configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuthInput"
      responses:
        "200":
          description: Passcode accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Incorrect passcode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin auth not configured on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login/reset:
    post:
      operationId: loginWithResetCode
      tags: [progress]
      summary: Redeem a one-time reset code to set a new password
      description: Verifies the player-supplied reset token and atomically sets a new password hash. The token is consumed on success. Score and progress are preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginResetInput"
      responses:
        "200":
          description: Reset accepted — progress returned with new password set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/{name}/password:
    delete:
      operationId: resetPlayerPassword
      tags: [progress]
      summary: Admin-initiated password reset
      description: Generates a one-time reset token for the named player. The token must be shared with the player so they can reclaim their account via POST /login/reset. Score, won keys, and times are preserved. Requires the SESSION_SECRET as admin passcode.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResetInput"
      responses:
        "200":
          description: Reset token generated — share it with the player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetResult"
        "401":
          description: Invalid admin passcode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin auth not configured on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /results/summary:
    get:
      operationId: getResultsSummary
      tags: [results]
      summary: Get leaderboard summary
      description: Returns aggregate leaderboard metrics for the admin dashboard.
      responses:
        "200":
          description: Leaderboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultsSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GameResult:
      type: object
      properties:
        id:
          type: number
        playerName:
          type: string
          description: Canonical (lowercased) player name — used as the key for password reset.
        firstName:
          type: string
        lastName:
          type: string
        score:
          type: number
        rank:
          type: string
        timeC1:
          type: ["number", "null"]
        timeC2:
          type: ["number", "null"]
        timeC3:
          type: ["number", "null"]
        timeC4:
          type: ["number", "null"]
        totalTime:
          type: ["number", "null"]
        isTest:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - playerName
        - firstName
        - lastName
        - score
        - rank
        - timeC1
        - timeC2
        - timeC3
        - timeC4
        - totalTime
        - isTest
        - createdAt
    GameResultInput:
      type: object
      properties:
        playerName:
          type: string
          minLength: 1
          description: The canonical player name (lowercased) used to look up the player record for authentication.
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        score:
          type: number
          minimum: 0
        rank:
          type: string
          minLength: 1
        timeC1:
          type: ["number", "null"]
        timeC2:
          type: ["number", "null"]
        timeC3:
          type: ["number", "null"]
        timeC4:
          type: ["number", "null"]
        totalTime:
          type: ["number", "null"]
        isTest:
          type: boolean
      required:
        - playerName
        - firstName
        - lastName
        - score
        - rank
        - timeC1
        - timeC2
        - timeC3
        - timeC4
        - totalTime
        - isTest
    ResultsSummary:
      type: object
      properties:
        participantCount:
          type: number
        averageScore:
          type: number
        topScore:
          type: ["number", "null"]
        averageTime:
          type: ["number", "null"]
        fastestTime:
          type: ["number", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - participantCount
        - averageScore
        - topScore
        - averageTime
        - fastestTime
        - updatedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Unique username (login key, lowercase-normalised).
        displayName:
          type: string
          description: Full display name shown on the leaderboard (optional on login, required on first registration).
        password:
          type: string
          minLength: 1
          format: password
      required:
        - name
        - password
    LoginResult:
      type: object
      properties:
        isNew:
          type: boolean
        name:
          type: string
        displayName:
          type: string
        score:
          type: number
        won:
          type: array
          items:
            type: number
        times:
          type: object
          additionalProperties: true
        submitted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - isNew
        - name
        - displayName
        - score
        - won
        - times
        - submitted
        - updatedAt
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - username
        - password
    AdminLoginResult:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
    AdminSessionResult:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
    ResetPasswordInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Username of the account to reset.
        password:
          type: string
          minLength: 1
          format: password
      required:
        - name
        - password
    PlayerProgress:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
        won:
          type: array
          items:
            type: number
        times:
          type: object
          additionalProperties: true
        submitted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - score
        - won
        - times
        - submitted
        - updatedAt
    AdminAuthInput:
      type: object
      properties:
        passcode:
          type: string
          minLength: 1
          description: The organizer passcode — validated against SESSION_SECRET on the server.
      required:
        - passcode
    AdminAuthResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    AdminResetInput:
      type: object
      properties:
        adminPasscode:
          type: string
          minLength: 1
          description: The organizer passcode — validated against SESSION_SECRET on the server.
      required:
        - adminPasscode
    PasswordResetResult:
      type: object
      description: Returned when an admin-initiated password reset succeeds. Contains the one-time token to share with the player.
      properties:
        name:
          type: string
        score:
          type: number
        won:
          type: array
          items:
            type: number
        times:
          type: object
          additionalProperties: true
        submitted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        resetToken:
          type: string
          description: One-time reset token — share this with the player. Valid for 24 hours.
        expiresAt:
          type: string
          format: date-time
          description: ISO timestamp when the reset token expires.
      required:
        - name
        - score
        - won
        - times
        - submitted
        - updatedAt
        - resetToken
        - expiresAt
    LoginResetInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        resetToken:
          type: string
          minLength: 1
          description: The one-time reset token issued by the admin.
        password:
          type: string
          minLength: 1
          format: password
      required:
        - name
        - resetToken
        - password
    PlayerProgressInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        score:
          type: number
          minimum: 0
        won:
          type: array
          items:
            type: number
        times:
          type: object
          additionalProperties: true
        submitted:
          type: boolean
      required:
        - name
        - score
        - won
        - times
        - submitted