# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /**
     * Unique username (login key, lowercase-normalised).
     * @minLength 1
     */
  name: string;
  /** Full display name shown on the leaderboard (optional on login, required on first registration). */
  displayName?: string;
  /** @minLength 1 */
  password: string;
}